[feat] border 재정의 로직 분리
